pyproject.toml
[project]
name = "llm-reasoning-activations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "datasets>=4.0.0",
    "matplotlib>=3.10.3",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "transformers>=4.53.2",
]
